@@ -14,8 +14,10 @@
 
 package org.skyscreamer.jsonassert;
 
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.skyscreamer.jsonassert.JSONCompareMode.LENIENT;
 import static org.skyscreamer.jsonassert.JSONCompareMode.NON_EXTENSIBLE;
 import static org.skyscreamer.jsonassert.JSONCompareMode.STRICT;
@@ -23,20 +25,20 @@ import static org.skyscreamer.jsonassert.JSONCompareMode.STRICT_ORDER;
 
 import java.util.Arrays;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.comparator.CustomComparator;
 import org.skyscreamer.jsonassert.comparator.JSONComparator;
+
+import com.github.openjson.JSONArray;
+import com.github.openjson.JSONException;
+import com.github.openjson.JSONObject;
 
 /**
  * Unit tests for {@link JSONAssert}
  */
-public class JSONAssertTest {
+class JSONAssertTest {
     @Test
-    public void testString() throws JSONException {
+    void testString() throws JSONException {
         testPass("\"Joe\"", "\"Joe\"", STRICT);
         testPass("\"Joe\"", "\"Joe\"", LENIENT);
         testPass("\"Joe\"", "\"Joe\"", NON_EXTENSIBLE);
@@ -48,7 +50,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testNumber() throws JSONException {
+    void testNumber() throws JSONException {
         testPass("123", "123", STRICT);
         testPass("123", "123", LENIENT);
         testPass("123", "123", NON_EXTENSIBLE);
@@ -66,7 +68,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testSimple() throws JSONException {
+    void testSimple() throws JSONException {
         testPass("{id:1}", "{id:1}", STRICT);
         testFail("{id:1}", "{id:2}", STRICT);
         testPass("{id:1}", "{id:1}", LENIENT);
@@ -78,7 +80,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testSimpleStrict() throws JSONException {
+    void testSimpleStrict() throws JSONException {
         testPass("{id:1}", "{id:1,name:\"Joe\"}", LENIENT);
         testFail("{id:1}", "{id:1,name:\"Joe\"}", STRICT);
         testPass("{id:1}", "{id:1,name:\"Joe\"}", STRICT_ORDER);
@@ -86,7 +88,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testReversed() throws JSONException {
+    void testReversed() throws JSONException {
         testPass("{name:\"Joe\",id:1}", "{id:1,name:\"Joe\"}", LENIENT);
         testPass("{name:\"Joe\",id:1}", "{id:1,name:\"Joe\"}", STRICT);
         testPass("{name:\"Joe\",id:1}", "{id:1,name:\"Joe\"}", NON_EXTENSIBLE);
@@ -94,7 +96,7 @@ public class JSONAssertTest {
     }
 
     @Test // Currently JSONAssert assumes JSONObject.
-    public void testArray() throws JSONException {
+    void testArray() throws JSONException {
         testPass("[1,2,3]","[1,2,3]", STRICT);
         testPass("[1,2,3]","[1,3,2]", LENIENT);
         testFail("[1,2,3]","[1,3,2]", STRICT);
@@ -106,7 +108,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testNested() throws JSONException {
+    void testNested() throws JSONException {
         testPass("{id:1,address:{addr1:\"123 Main\", addr2:null, city:\"Houston\", state:\"TX\"}}",
                 "{id:1,address:{addr1:\"123 Main\", addr2:null, city:\"Houston\", state:\"TX\"}}", STRICT);
         testFail("{id:1,address:{addr1:\"123 Main\", addr2:null, city:\"Houston\", state:\"TX\"}}",
@@ -114,7 +116,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testVeryNested() throws JSONException {
+    void testVeryNested() throws JSONException {
         testPass("{a:{b:{c:{d:{e:{f:{g:{h:{i:{j:{k:{l:{m:{n:{o:{p:\"blah\"}}}}}}}}}}}}}}}}",
                 "{a:{b:{c:{d:{e:{f:{g:{h:{i:{j:{k:{l:{m:{n:{o:{p:\"blah\"}}}}}}}}}}}}}}}}", STRICT);
         testFail("{a:{b:{c:{d:{e:{f:{g:{h:{i:{j:{k:{l:{m:{n:{o:{p:\"blah\"}}}}}}}}}}}}}}}}",
@@ -122,7 +124,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testSimpleArray() throws JSONException {
+    void testSimpleArray() throws JSONException {
         testPass("{id:1,pets:[\"dog\",\"cat\",\"fish\"]}", // Exact to exact (strict)
                 "{id:1,pets:[\"dog\",\"cat\",\"fish\"]}",
                 STRICT);
@@ -153,40 +155,40 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testSimpleMixedArray() throws JSONException {
+    void testSimpleMixedArray() throws JSONException {
         testPass("{stuff:[321, \"abc\"]}", "{stuff:[\"abc\", 321]}", LENIENT);
         testFail("{stuff:[321, \"abc\"]}", "{stuff:[\"abc\", 789]}", LENIENT);
     }
 
     @Test
-    public void testComplexMixedStrictArray() throws JSONException {
+    void testComplexMixedStrictArray() throws JSONException {
         testPass("{stuff:[{pet:\"cat\"},{car:\"Ford\"}]}", "{stuff:[{pet:\"cat\"},{car:\"Ford\"}]}", STRICT);
     }
 
     @Test
-    public void testComplexMixedArray() throws JSONException {
+    void testComplexMixedArray() throws JSONException {
         testPass("{stuff:[{pet:\"cat\"},{car:\"Ford\"}]}", "{stuff:[{pet:\"cat\"},{car:\"Ford\"}]}", LENIENT);
     }
 
     @Test
-    public void testComplexArrayNoUniqueID() throws JSONException {
+    void testComplexArrayNoUniqueID() throws JSONException {
         testPass("{stuff:[{address:{addr1:\"123 Main\"}}, {address:{addr1:\"234 Broad\"}}]}",
                 "{stuff:[{address:{addr1:\"123 Main\"}}, {address:{addr1:\"234 Broad\"}}]}",
                 LENIENT);
     }
 
     @Test
-    public void testSimpleAndComplexStrictArray() throws JSONException {
+    void testSimpleAndComplexStrictArray() throws JSONException {
         testPass("{stuff:[123,{a:\"b\"}]}", "{stuff:[123,{a:\"b\"}]}", STRICT);
     }
 
     @Test
-    public void testSimpleAndComplexArray() throws JSONException {
+    void testSimpleAndComplexArray() throws JSONException {
         testPass("{stuff:[123,{a:\"b\"}]}", "{stuff:[123,{a:\"b\"}]}", LENIENT);
     }
 
     @Test
-    public void testComplexArray() throws JSONException {
+    void testComplexArray() throws JSONException {
         testPass("{id:1,name:\"Joe\",friends:[{id:2,name:\"Pat\",pets:[\"dog\"]},{id:3,name:\"Sue\",pets:[\"bird\",\"fish\"]}],pets:[]}",
                  "{id:1,name:\"Joe\",friends:[{id:2,name:\"Pat\",pets:[\"dog\"]},{id:3,name:\"Sue\",pets:[\"bird\",\"fish\"]}],pets:[]}",
                  STRICT); // Exact to exact (strict)
@@ -217,32 +219,32 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testArrayOfArraysStrict() throws JSONException {
+    void testArrayOfArraysStrict() throws JSONException {
         testPass("{id:1,stuff:[[1,2],[2,3],[],[3,4]]}", "{id:1,stuff:[[1,2],[2,3],[],[3,4]]}", STRICT);
         testFail("{id:1,stuff:[[1,2],[2,3],[3,4],[]]}", "{id:1,stuff:[[1,2],[2,3],[],[3,4]]}", STRICT);
     }
 
     @Test
-    public void testArrayOfArrays() throws JSONException {
+    void testArrayOfArrays() throws JSONException {
         testPass("{id:1,stuff:[[4,3],[3,2],[],[1,2]]}", "{id:1,stuff:[[1,2],[2,3],[],[3,4]]}", LENIENT);
-    }
-    
-    @Test
-    public void testLenientArrayRecursion() throws JSONException {
-        testPass("[{\"arr\":[5, 2, 1]}]", "[{\"b\":3, \"arr\":[1, 5, 2]}]", LENIENT);
-    }
-   
-    @Test 
-    public void testFieldMismatch() throws JSONException {
-        JSONCompareResult result = JSONCompare.compareJSON("{name:\"Pat\"}", "{name:\"Sue\"}", STRICT);
-        FieldComparisonFailure comparisonFailure = result.getFieldFailures().iterator().next();
-        Assert.assertEquals("Pat", comparisonFailure.getExpected());
-        Assert.assertEquals("Sue", comparisonFailure.getActual());
-        Assert.assertEquals("name", comparisonFailure.getField());
     }
 
     @Test
-    public void testBooleanArray() throws JSONException {
+    void testLenientArrayRecursion() throws JSONException {
+        testPass("[{\"arr\":[5, 2, 1]}]", "[{\"b\":3, \"arr\":[1, 5, 2]}]", LENIENT);
+    }
+
+    @Test
+    void testFieldMismatch() throws JSONException {
+        JSONCompareResult result = JSONCompare.compareJSON("{name:\"Pat\"}", "{name:\"Sue\"}", STRICT);
+        FieldComparisonFailure comparisonFailure = result.getFieldFailures().iterator().next();
+        assertEquals("Pat", comparisonFailure.getExpected());
+        assertEquals("Sue", comparisonFailure.getActual());
+        assertEquals("name", comparisonFailure.getField());
+    }
+
+    @Test
+    void testBooleanArray() throws JSONException {
         testPass("[true, false, true, true, false]", "[true, false, true, true, false]", STRICT);
         testPass("[false, true, true, false, true]", "[true, false, true, true, false]", LENIENT);
         testFail("[false, true, true, false, true]", "[true, false, true, true, false]", STRICT);
@@ -251,34 +253,34 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testNullProperty() throws JSONException {
+    void testNullProperty() throws JSONException {
         testFail("{id:1,name:\"Joe\"}", "{id:1,name:null}", STRICT);
         testFail("{id:1,name:null}", "{id:1,name:\"Joe\"}", STRICT);
     }
 
     @Test
-    public void testIncorrectTypes() throws JSONException {
+    void testIncorrectTypes() throws JSONException {
         testFail("{id:1,name:\"Joe\"}", "{id:1,name:[]}", STRICT);
         testFail("{id:1,name:[]}", "{id:1,name:\"Joe\"}", STRICT);
     }
 
     @Test
-    public void testNullEquality() throws JSONException {
+    void testNullEquality() throws JSONException {
         testPass("{id:1,name:null}", "{id:1,name:null}", STRICT);
     }
 
     @Test
-    public void testExpectedArrayButActualObject() throws JSONException {
+    void testExpectedArrayButActualObject() throws JSONException {
         testFail("[1]", "{id:1}", LENIENT);
     }
 
     @Test
-    public void testExpectedObjectButActualArray() throws JSONException {
+    void testExpectedObjectButActualArray() throws JSONException {
         testFail("{id:1}", "[1]", LENIENT);
     }
 
     @Test
-    public void testEquivalentIntAndLong() throws JSONException {
+    void testEquivalentIntAndLong() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
@@ -288,7 +290,7 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testEquivalentIntAndDouble() throws JSONException {
+    void testEquivalentIntAndDouble() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
@@ -297,26 +299,26 @@ public class JSONAssertTest {
         JSONAssert.assertEquals(actual, expected, true);
     }
 
-    @Test(expected = AssertionError.class)
-    public void testAssertNotEqualsWhenEqualStrict() throws JSONException {
+    @Test
+    void testAssertNotEqualsWhenEqualStrict() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("id", Double.valueOf(12345));
-        JSONAssert.assertNotEquals(expected, actual, true);
+        assertThrows(AssertionError.class, () -> JSONAssert.assertNotEquals(expected, actual, true));
     }
 
-    @Test(expected = AssertionError.class)
-    public void testAssertNotEqualsWhenEqualLenient() throws JSONException {
+    @Test
+    void testAssertNotEqualsWhenEqualLenient() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("id", Double.valueOf(12345));
-        JSONAssert.assertNotEquals(expected, actual, false);
+        assertThrows(AssertionError.class, () -> JSONAssert.assertNotEquals(expected, actual, false));
     }
 
     @Test()
-    public void testAssertNotEqualsWhenEqualDiffObjectsStrict() throws JSONException {
+    void testAssertNotEqualsWhenEqualDiffObjectsStrict() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
@@ -325,19 +327,19 @@ public class JSONAssertTest {
         JSONAssert.assertNotEquals(expected, actual, true);
     }
 
-    @Test(expected = AssertionError.class)
-    public void testAssertNotEqualsWhenEqualDiffObjectsLenient() throws JSONException {
+    @Test
+    void testAssertNotEqualsWhenEqualDiffObjectsLenient() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         expected.put("name", "Joe");
         actual.put("name", "Joe");
         actual.put("id", Double.valueOf(12345));
-        JSONAssert.assertNotEquals(expected, actual, false);
+        assertThrows(AssertionError.class, () -> JSONAssert.assertNotEquals(expected, actual, false));
     }
 
     @Test()
-    public void testAssertNotEqualsWhenDifferentStrict() throws JSONException {
+    void testAssertNotEqualsWhenDifferentStrict() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
@@ -346,7 +348,7 @@ public class JSONAssertTest {
     }
 
     @Test()
-    public void testAssertNotEqualsWhenDifferentLenient() throws JSONException {
+    void testAssertNotEqualsWhenDifferentLenient() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
@@ -355,15 +357,15 @@ public class JSONAssertTest {
     }
 
     @Test()
-    public void testAssertNotEqualsString() throws JSONException {
+    void testAssertNotEqualsString() throws JSONException {
         JSONAssert.assertNotEquals("[1,2,3]", "[1,3,2]", STRICT);
         JSONAssert.assertNotEquals("[1,2,3]", "[1,2,4]", LENIENT);
         JSONAssert.assertNotEquals("[1,2,3]", "[1,3,2]", true);
         JSONAssert.assertNotEquals("[1,2,3]", "[1,2,4]", false);
     }
-    
+
     @Test()
-    public void testAssertEqualsString() throws JSONException {
+    void testAssertEqualsString() throws JSONException {
         JSONAssert.assertEquals("[1,2,3]", "[1,2,3]", true);
         JSONAssert.assertEquals("{id:12345}", "{id:12345}", false);
         JSONAssert.assertEquals("{id:12345}", "{id:12345, name:\"john\"}", LENIENT);
@@ -372,7 +374,7 @@ public class JSONAssertTest {
     }
 
     @Test()
-    public void testAssertNotEqualsStringAndJSONObject() throws JSONException {
+    void testAssertNotEqualsStringAndJSONObject() throws JSONException {
         JSONObject actual = new JSONObject();
         actual.put("id", Double.valueOf(12345));
         JSONAssert.assertEquals("{id:12345}", actual, false);
@@ -380,7 +382,7 @@ public class JSONAssertTest {
     }
 
     @Test()
-    public void testAssertNotEqualsJSONArray() throws JSONException {
+    void testAssertNotEqualsJSONArray() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertEquals("[1,2,3]", actual, false);
         JSONAssert.assertNotEquals("[1,2,4]", actual, false);
@@ -388,17 +390,17 @@ public class JSONAssertTest {
         JSONAssert.assertNotEquals(new JSONArray(Arrays.asList(1, 2, 4)), actual, false);
         JSONAssert.assertNotEquals(new JSONArray(Arrays.asList(1, 3, 2)), actual, true);
     }
-    
+
     @Test
-    public void testAssertEqualsStringJSONArrayBooleanWithMessage() throws JSONException {
+    void testAssertEqualsStringJSONArrayBooleanWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertEquals("Message", "[1,2,3]", actual, false);
         performAssertEqualsTestForMessageVerification("[1,2,4]", actual, false);
         performAssertEqualsTestForMessageVerification("[1,3,2]", actual, true);
     }
-    
+
     @Test
-    public void testAssertEqualsStringJSONArrayCompareModeWithMessage() throws JSONException {
+    void testAssertEqualsStringJSONArrayCompareModeWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertEquals("Message", "[1,2,3]", actual, LENIENT);
         performAssertEqualsTestForMessageVerification("[1,2,4]", actual, LENIENT);
@@ -406,247 +408,247 @@ public class JSONAssertTest {
     }
 
     @Test
-    public void testAssertEqualsJSONArray2BooleanWithMessage() throws JSONException {
+    void testAssertEqualsJSONArray2BooleanWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertEquals("Message", new JSONArray(Arrays.asList(1, 2, 3)), actual, false);
         performAssertEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 2, 4)), actual, false);
         performAssertEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 3, 2)), actual, true);
     }
-    
+
     @Test
-    public void testAssertEqualsJSONArray2JSONCompareWithMessage() throws JSONException {
+    void testAssertEqualsJSONArray2JSONCompareWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
-        
+
         JSONAssert.assertEquals("Message", new JSONArray(Arrays.asList(1, 2, 3)), actual, LENIENT);
         performAssertEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 2, 4)), actual, LENIENT);
         performAssertEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 3, 2)), actual, STRICT);
     }
-    
+
     @Test
-    public void testAssertEqualsString2Boolean() throws JSONException {
+    void testAssertEqualsString2Boolean() throws JSONException {
         JSONAssert.assertEquals("Message", "{id:12345}", "{id:12345}", false);
         JSONAssert.assertEquals("Message", "{id:12345}", "{id:12345, name:\"john\"}", false);
-        
+
         performAssertEqualsTestForMessageVerification("{id:12345}", "{id:12345, name:\"john\"}", true);
         performAssertEqualsTestForMessageVerification("{id:12345}", "{id:123456}", false);
     }
-    
+
     @Test
-    public void testAssertEqualsString2JSONCompare() throws JSONException {
+    void testAssertEqualsString2JSONCompare() throws JSONException {
         JSONAssert.assertEquals("Message", "{id:12345}", "{id:12345}", LENIENT);
         JSONAssert.assertEquals("Message", "{id:12345}", "{id:12345, name:\"john\"}", LENIENT);
-        
+
         performAssertEqualsTestForMessageVerification("{id:12345}", "{id:12345, name:\"john\"}", STRICT);
         performAssertEqualsTestForMessageVerification("{id:12345}", "{id:123456}", LENIENT);
     }
-    
+
     @Test
-    public void testAssertEqualsStringJSONObjectBoolean() throws JSONException {
+    void testAssertEqualsStringJSONObjectBoolean() throws JSONException {
         JSONObject actual = new JSONObject();
         actual.put("id", Double.valueOf(12345));
         JSONAssert.assertEquals("Message", "{id:12345}", actual, false);
         performAssertEqualsTestForMessageVerification("{id:12346}", actual, false);
         performAssertEqualsTestForMessageVerification("[1,2,3]", "[1,3,2]", true);
     }
-    
+
     @Test
-    public void testAssertEqualsStringJSONObjectJSONCompare() throws JSONException {
+    void testAssertEqualsStringJSONObjectJSONCompare() throws JSONException {
         JSONObject actual = new JSONObject();
         actual.put("id", Double.valueOf(12345));
         JSONAssert.assertEquals("Message", "{id:12345}", actual, LENIENT);
         performAssertEqualsTestForMessageVerification("{id:12346}", actual, LENIENT);
         performAssertEqualsTestForMessageVerification("[1,2,3]", "[1,3,2]", STRICT);
     }
-    
+
     @Test
-    public void testAssertEqualsJSONObject2JSONCompare() throws JSONException {
+    void testAssertEqualsJSONObject2JSONCompare() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("name", "Joe");
         actual.put("id", Integer.valueOf(12345));
         JSONAssert.assertEquals("Message", expected, actual, LENIENT);
-        
+
         expected.put("street", "St. Paul");
         performAssertEqualsTestForMessageVerification(expected, actual, LENIENT);
-        
+
         expected = new JSONObject();
         actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("id", Double.valueOf(12346));
         performAssertEqualsTestForMessageVerification(expected, actual, STRICT);
     }
-    
+
     @Test
-    public void testAssertEqualsJSONObject2Boolean() throws JSONException {
+    void testAssertEqualsJSONObject2Boolean() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("name", "Joe");
         actual.put("id", Integer.valueOf(12345));
         JSONAssert.assertEquals("Message", expected, actual, false);
-        
+
         expected.put("street", "St. Paul");
         performAssertEqualsTestForMessageVerification(expected, actual, false);
-        
+
         expected = new JSONObject();
         actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("id", Double.valueOf(12346));
         performAssertEqualsTestForMessageVerification(expected, actual, true);
     }
-    
+
     @Test
-    public void testAssertEqualsString2JsonComparator() throws IllegalArgumentException, JSONException {
-        JSONAssert.assertEquals("Message", "{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":2}}", 
+    void testAssertEqualsString2JsonComparator() throws IllegalArgumentException, JSONException {
+        JSONAssert.assertEquals("Message", "{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":2}}",
             new CustomComparator(
-                JSONCompareMode.STRICT, 
-                new Customization("entry.id", 
+                JSONCompareMode.STRICT,
+                new Customization("entry.id",
                 new RegularExpressionValueMatcher<Object>("\\d"))
          ));
-        
-        performAssertEqualsTestForMessageVerification("{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":as}}", 
+
+        performAssertEqualsTestForMessageVerification("{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":as}}",
             new CustomComparator(
-                JSONCompareMode.STRICT, 
-                new Customization("entry.id", 
+                JSONCompareMode.STRICT,
+                new Customization("entry.id",
                 new RegularExpressionValueMatcher<Object>("\\d"))
         ));
     }
-    
+
     @Test
-    public void testAssertNotEqualsStringJSONArrayBooleanWithMessage() throws JSONException {
+    void testAssertNotEqualsStringJSONArrayBooleanWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertNotEquals("Message", "[1,4,3]", actual, false);
         JSONAssert.assertNotEquals("Message", "[1,4,3]", actual, true);
         performAssertNotEqualsTestForMessageVerification("[1,3,2]", actual, false);
         performAssertNotEqualsTestForMessageVerification("[1,2,3]", actual, true);
     }
-    
+
     @Test
-    public void testAssertNotEqualsStringJSONArrayCompareModeWithMessage() throws JSONException {
+    void testAssertNotEqualsStringJSONArrayCompareModeWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertNotEquals("Message", "[1,2,4]", actual, LENIENT);
         JSONAssert.assertNotEquals("Message", "[1,2,4]", actual, STRICT);
         performAssertNotEqualsTestForMessageVerification("[1,3,2]", actual, LENIENT);
         performAssertNotEqualsTestForMessageVerification("[1,2,3]", actual, STRICT);
     }
-    
+
     @Test
-    public void testAssertNotEqualsJSONArray2BooleanWithMessage() throws JSONException {
+    void testAssertNotEqualsJSONArray2BooleanWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
         JSONAssert.assertNotEquals("Message", new JSONArray(Arrays.asList(1, 4, 3)), actual, false);
         performAssertNotEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 3, 2)), actual, false);
         performAssertNotEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 2, 3)), actual, true);
     }
-    
+
     @Test
-    public void testAssertNotEqualsJSONArray2JSONCompareWithMessage() throws JSONException {
+    void testAssertNotEqualsJSONArray2JSONCompareWithMessage() throws JSONException {
         JSONArray actual = new JSONArray(Arrays.asList(1, 2, 3));
-        
+
         JSONAssert.assertNotEquals("Message", new JSONArray(Arrays.asList(1, 4, 3)), actual, LENIENT);
         performAssertNotEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 3, 2)), actual, LENIENT);
         performAssertNotEqualsTestForMessageVerification(new JSONArray(Arrays.asList(1, 2, 3)), actual, STRICT);
     }
-    
+
     @Test
-    public void testAssertNotEqualsString2Boolean() throws JSONException {
+    void testAssertNotEqualsString2Boolean() throws JSONException {
         JSONAssert.assertNotEquals("Message", "{id:12345}", "{id:45}", false);
         JSONAssert.assertNotEquals("Message", "{id:12345}", "{id:345, name:\"john\"}", false);
-        
+
         performAssertNotEqualsTestForMessageVerification("{id:12345}", "{id:12345}", true);
         performAssertNotEqualsTestForMessageVerification("{id:12345}", "{id:12345, name:\"John\"}", false);
     }
-    
+
     @Test
-    public void testAssertNotEqualsString2JSONCompare() throws JSONException {
+    void testAssertNotEqualsString2JSONCompare() throws JSONException {
         JSONAssert.assertNotEquals("Message", "{id:12345}", "{id:123}", LENIENT);
         JSONAssert.assertNotEquals("Message", "{id:12345, name:\"John\"}", "{id:12345}", LENIENT);
-        
+
         performAssertNotEqualsTestForMessageVerification("{id:12345}", "{id:12345, name:\"john\"}", LENIENT);
         performAssertNotEqualsTestForMessageVerification("{id:12345}", "{id:12345}", STRICT);
     }
-    
+
     @Test
-    public void testAssertNotEqualsStringJSONObjectBoolean() throws JSONException {
+    void testAssertNotEqualsStringJSONObjectBoolean() throws JSONException {
         JSONObject actual = new JSONObject();
         actual.put("id", Double.valueOf(12345));
         JSONAssert.assertNotEquals("Message", "{id:1234}", actual, false);
         performAssertNotEqualsTestForMessageVerification("{id:12345}", actual, false);
         performAssertNotEqualsTestForMessageVerification("[1,2,3]", "[1,2,3]", true);
     }
-    
+
     @Test
-    public void testAssertNotEqualsStringJSONObjectJSONCompare() throws JSONException {
+    void testAssertNotEqualsStringJSONObjectJSONCompare() throws JSONException {
         JSONObject actual = new JSONObject();
         actual.put("id", Double.valueOf(12345));
         JSONAssert.assertNotEquals("Message", "{id:1234}", actual, LENIENT);
         performAssertNotEqualsTestForMessageVerification("{id:12345}", actual, LENIENT);
         performAssertNotEqualsTestForMessageVerification("[1,2,3]", "[1,2,3]", STRICT);
     }
-    
+
     @Test
-    public void testAssertNtEqualsJSONObject2JSONCompare() throws JSONException {
+    void testAssertNtEqualsJSONObject2JSONCompare() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("name", "Joe");
         actual.put("id", Integer.valueOf(123));
         JSONAssert.assertNotEquals("Message", expected, actual, LENIENT);
-        
+
         actual.remove("id");
         actual.put("id", Integer.valueOf(12345));
         performAssertNotEqualsTestForMessageVerification(expected, actual, LENIENT);
-        
+
         expected = new JSONObject();
         actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("id", Double.valueOf(12345));
         performAssertNotEqualsTestForMessageVerification(expected, actual, STRICT);
     }
-    
+
     @Test
-    public void testAssertNotEqualsJSONObject2Boolean() throws JSONException {
+    void testAssertNotEqualsJSONObject2Boolean() throws JSONException {
         JSONObject expected = new JSONObject();
         JSONObject actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("name", "Joe");
         actual.put("id", Integer.valueOf(123));
         JSONAssert.assertNotEquals("Message", expected, actual, false);
-        
+
         actual.remove("id");
         actual.put("id", Integer.valueOf(12345));
         performAssertNotEqualsTestForMessageVerification(expected, actual, false);
-        
+
         expected = new JSONObject();
         actual = new JSONObject();
         expected.put("id", Integer.valueOf(12345));
         actual.put("id", Double.valueOf(12345));
         performAssertNotEqualsTestForMessageVerification(expected, actual, true);
     }
-    
+
     @Test
-    public void testAssertNotEqualsString2JsonComparator() throws IllegalArgumentException, JSONException {
-        JSONAssert.assertNotEquals("Message", "{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":hh}}", 
+    void testAssertNotEqualsString2JsonComparator() throws IllegalArgumentException, JSONException {
+        JSONAssert.assertNotEquals("Message", "{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":hh}}",
             new CustomComparator(
-                JSONCompareMode.STRICT, 
-                new Customization("entry.id", 
+                JSONCompareMode.STRICT,
+                new Customization("entry.id",
                 new RegularExpressionValueMatcher<Object>("\\d"))
          ));
-        
-        performAssertNotEqualsTestForMessageVerification("{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":2}}", 
+
+        performAssertNotEqualsTestForMessageVerification("{\"entry\":{\"id\":x}}", "{\"entry\":{\"id\":1, \"id\":2}}",
             new CustomComparator(
-                JSONCompareMode.STRICT, 
-                new Customization("entry.id", 
+                JSONCompareMode.STRICT,
+                new Customization("entry.id",
                 new RegularExpressionValueMatcher<Object>("\\d"))
         ));
     }
-    
+
     private void testPass(String expected, String actual, JSONCompareMode compareMode)
             throws JSONException
     {
         String message = expected + " == " + actual + " (" + compareMode + ")";
         JSONCompareResult result = JSONCompare.compareJSON(expected, actual, compareMode);
-        Assert.assertTrue(message + "\n  " + result.getMessage(), result.passed());
+        assertTrue(result.passed(), message + "\n  " + result.getMessage());
     }
 
     private void testFail(String expected, String actual, JSONCompareMode compareMode)
@@ -654,14 +656,14 @@ public class JSONAssertTest {
     {
         String message = expected + " != " + actual + " (" + compareMode + ")";
         JSONCompareResult result = JSONCompare.compareJSON(expected, actual, compareMode);
-        Assert.assertTrue(message, result.failed());
+        assertTrue(result.failed(), message);
     }
-    
+
     private void performAssertEqualsTestForMessageVerification(
-        Object expected, 
-        Object actual, 
+        Object expected,
+        Object actual,
         Object strictMode) throws JSONException {
-        
+
         String message = "Message";
         String testShouldFailMessage = "The test should fail so that the message in AssertionError could be verified.";
         String strictModeMessage = "strictMode must be an instance of JSONCompareMode or Boolean";
@@ -749,13 +751,13 @@ public class JSONAssertTest {
             fail("No overloaded method found to call");
         }
     }
-    
+
     private void performAssertNotEqualsTestForMessageVerification(
-        Object expected, 
-        Object actual, 
-        Object strictMode) 
+        Object expected,
+        Object actual,
+        Object strictMode)
         throws JSONException {
-        
+
         String message = "Message";
         String testShouldFailMessage = "The test should fail so that the message in AssertionError could be verified.";
         String strictModeMessage = "strictMode must be an instance of JSONCompareMode or Boolean";
@@ -851,7 +853,7 @@ public class JSONAssertTest {
             throw ae;
         }
     }
-    
+
     private void verifyErrorMessage(String message, AssertionError ae) {
         assertTrue(ae.getMessage().contains(message));
         assertTrue(ae.getMessage().startsWith(message));
