@@ -14,21 +14,20 @@
 
 package org.skyscreamer.jsonassert;
 
-import org.junit.Assert;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.json.JSONException;
-import org.junit.Test;
-import org.skyscreamer.jsonassert.Customization;
-import org.skyscreamer.jsonassert.JSONCompareMode;
+import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.comparator.CustomComparator;
+
+import com.github.openjson.JSONException;
 
 /**
  * Unit tests for RegularExpressionValueMatcher
- * 
+ *
  * @author Duncan Mackinder
- * 
+ *
  */
-public class RegularExpressionValueMatcherTest {
+class RegularExpressionValueMatcherTest {
 	private static final String ARRAY_ELEMENT_PREFIX = "d.results[0].__metadata.uri";
 	private static final String JSON_STRING_WITH_ARRAY = "{d:{results:[{__metadata:{uri:\"http://localhost:80/Person('1')\",type:Person},id:1}]}}";
 	private static final String CONSTANT_URI_REGEX_EXPECTED_JSON = "{d:{results:[{__metadata:{uri:X}}]}}";
@@ -41,80 +40,80 @@ public class RegularExpressionValueMatcherTest {
 	}
 
 	@Test
-	public void constantRegexWithSimplePathMatchsStringAttribute() throws JSONException {
+	void constantRegexWithSimplePathMatchsStringAttribute() throws JSONException {
 		doTest("a", "v.", "{a:x}", "{a:v1}");
 	}
 
 	@Test
-	public void constantRegexWithThreeLevelPathMatchsStringAttribute() throws JSONException {
+	void constantRegexWithThreeLevelPathMatchsStringAttribute() throws JSONException {
 		doTest("a.b.c", ".*Is.*", "{a:{b:{c:x}}}", "{a:{b:{c:thisIsAString}}}");
 	}
 
 	@Test
-	public void dynamicRegexWithSimplePathMatchsStringAttribute() throws JSONException {
+	void dynamicRegexWithSimplePathMatchsStringAttribute() throws JSONException {
 		doTest("a", null, "{a:\"v.\"}", "{a:v1}");
 	}
 
 	@Test
-	public void dynamicRegexWithThreeLevelPathMatchsStringAttribute() throws JSONException {
+	void dynamicRegexWithThreeLevelPathMatchsStringAttribute() throws JSONException {
 		doTest("a.b.c", null, "{a:{b:{c:\".*Is.*\"}}}",
 				"{a:{b:{c:thisIsAString}}}");
 	}
 
 	@Test
-	public void constantRegexMatchesStringAttributeInsideArray() throws JSONException {
+	void constantRegexMatchesStringAttributeInsideArray() throws JSONException {
 		doTest(ARRAY_ELEMENT_PREFIX, "http://localhost:80/Person\\('\\d+'\\)", CONSTANT_URI_REGEX_EXPECTED_JSON, JSON_STRING_WITH_ARRAY);
 	}
 
     @Test
-    public void dynamicRegexMatchesStringAttributeInsideArray() throws JSONException {
+    void dynamicRegexMatchesStringAttributeInsideArray() throws JSONException {
         doTest(ARRAY_ELEMENT_PREFIX, null, "{d:{results:[{__metadata:{uri:\"http://localhost:80/Person\\\\('\\\\d+'\\\\)\"}}]}}", JSON_STRING_WITH_ARRAY);
     }
 
     @Test
-    public void dynamicRegexMatchesStringAttributeInsideArrayWithNoArgConstructor() throws JSONException {
+    void dynamicRegexMatchesStringAttributeInsideArrayWithNoArgConstructor() throws JSONException {
 		JSONAssert.assertEquals("{d:{results:[{__metadata:{uri:\"http://localhost:80/Person\\\\('\\\\d+'\\\\)\"}}]}}", JSON_STRING_WITH_ARRAY, new CustomComparator(
 				JSONCompareMode.STRICT_ORDER, new Customization(ARRAY_ELEMENT_PREFIX,
 						new RegularExpressionValueMatcher<Object>())));
     }
 
     @Test
-    public void failsWhenDynamicRegexInvalid() throws JSONException {
+    void failsWhenDynamicRegexInvalid() throws JSONException {
     	try {
     		doTest(ARRAY_ELEMENT_PREFIX, null, "{d:{results:[{__metadata:{uri:\"http://localhost:80/Person('\\\\d+'\\\\)\"}}]}}", JSON_STRING_WITH_ARRAY);
     	}
     	catch (AssertionError e) {
-    		Assert.assertTrue("Invalid exception message returned: "+ e.getMessage(), e.getMessage().startsWith(ARRAY_ELEMENT_PREFIX + ": Dynamic expected pattern invalid: "));
+    		assertTrue(e.getMessage().startsWith(ARRAY_ELEMENT_PREFIX + ": Dynamic expected pattern invalid: "), "Invalid exception message returned: "+ e.getMessage());
     	}
     }
 
     @Test
-    public void failsWhenDynamicRegexDoesNotMatchStringAttributeInsideArray() throws JSONException {
+    void failsWhenDynamicRegexDoesNotMatchStringAttributeInsideArray() throws JSONException {
     	try {
     		doTest(ARRAY_ELEMENT_PREFIX, null, "{d:{results:[{__metadata:{uri:\"http://localhost:80/Person\\\\('\\\\w+'\\\\)\"}}]}}", JSON_STRING_WITH_ARRAY);
     	}
     	catch (AssertionError e) {
-    		Assert.assertTrue("Invalid exception message returned: "+ e.getMessage(), e.getMessage().startsWith(ARRAY_ELEMENT_PREFIX + ": Dynamic expected pattern did not match value"));
+    		assertTrue(e.getMessage().startsWith(ARRAY_ELEMENT_PREFIX + ": Dynamic expected pattern did not match value"), "Invalid exception message returned: "+ e.getMessage());
     	}
     }
 
     @Test
-    public void failsWhenConstantRegexInvalid() throws JSONException {
+    void failsWhenConstantRegexInvalid() throws JSONException {
     	try {
     		doTest(ARRAY_ELEMENT_PREFIX, "http://localhost:80/Person\\\\['\\\\d+'\\\\)", CONSTANT_URI_REGEX_EXPECTED_JSON, JSON_STRING_WITH_ARRAY);
     	}
     	catch (IllegalArgumentException e) {
-    		Assert.assertTrue("Invalid exception message returned: "+ e.getMessage(), e.getMessage().startsWith("Constant expected pattern invalid: "));
+    		assertTrue(e.getMessage().startsWith("Constant expected pattern invalid: "), "Invalid exception message returned: "+ e.getMessage());
     	}
     }
 
     @Test
-    public void failsWhenConstantRegexDoesNotMatchStringAttributeInsideArray() throws JSONException {
+    void failsWhenConstantRegexDoesNotMatchStringAttributeInsideArray() throws JSONException {
     	try {
     		doTest(ARRAY_ELEMENT_PREFIX, "http://localhost:80/Person\\\\('\\\\w+'\\\\)", CONSTANT_URI_REGEX_EXPECTED_JSON, JSON_STRING_WITH_ARRAY);
     	}
     	catch (AssertionError e) {
-    		Assert.assertTrue("Invalid exception message returned: "+ e.getMessage(), e.getMessage().startsWith(ARRAY_ELEMENT_PREFIX + ": Constant expected pattern did not match value"));
+    		assertTrue(e.getMessage().startsWith(ARRAY_ELEMENT_PREFIX + ": Constant expected pattern did not match value"), "Invalid exception message returned: "+ e.getMessage());
     	}
     }
 }
