@@ -16,9 +16,10 @@ package org.skyscreamer.jsonassert;
 
 import java.text.MessageFormat;
 
-import org.json.JSONArray;
-import org.json.JSONException;
 import org.skyscreamer.jsonassert.comparator.JSONComparator;
+
+import com.github.openjson.JSONArray;
+import com.github.openjson.JSONException;
 
 /**
  * <p>A value matcher for arrays. This operates like STRICT_ORDER array match,
@@ -31,22 +32,22 @@ import org.skyscreamer.jsonassert.comparator.JSONComparator;
  * array element components against a regular expression pattern. As a convenience to reduce syntactic complexity of expected string, if the
  * expected object is not an array, a one element expected array is created
  * containing whatever is provided as the expected value.</p>
- * 
+ *
  * <p>Some examples of typical usage idioms listed below.</p>
- * 
+ *
  * <p>Assuming JSON to be verified is held in String variable ARRAY_OF_JSONOBJECTS and contains:</p>
- * 
+ *
  * <pre>{@code
  * {a:[{background:white, id:1, type:row},
  *     {background:grey,  id:2, type:row},
  *     {background:white, id:3, type:row},
  *     {background:grey,  id:4, type:row}]}
  * }</pre>
- * 
+ *
  * <p>then:</p>
- * 
+ *
  * <p>To verify that the 'id' attribute of first element of array 'a' is '1':</p>
- * 
+ *
  * <pre>{@code
  * JSONComparator comparator = new DefaultComparator(JSONCompareMode.LENIENT);
  * Customization customization = new Customization("a", new ArrayValueMatcher<Object>(comparator, 0));
@@ -55,27 +56,27 @@ import org.skyscreamer.jsonassert.comparator.JSONComparator;
  * }</pre>
  *
  * <p>To simplify complexity of expected JSON string, the value <code>"a:[{id:1}]}"</code> may be replaced by <code>"a:{id:1}}"</code></p>
- * 
+ *
  * <p>To verify that the 'type' attribute of second and third elements of array 'a' is 'row':</p>
- * 
+ *
  * <pre>{@code
  * JSONComparator comparator = new DefaultComparator(JSONCompareMode.LENIENT);
  * Customization customization = new Customization("a", new ArrayValueMatcher<Object>(comparator, 1, 2));
  * JSONAssert.assertEquals("{a:[{type:row}]}", ARRAY_OF_JSONOBJECTS,
  *     new CustomComparator(JSONCompareMode.LENIENT, customization));
  * }</pre>
- * 
+ *
  * <p>To verify that the 'type' attribute of every element of array 'a' is 'row':</p>
- * 
+ *
  * <pre>{@code
  * JSONComparator comparator = new DefaultComparator(JSONCompareMode.LENIENT);
  * Customization customization = new Customization("a", new ArrayValueMatcher<Object>(comparator));
  * JSONAssert.assertEquals("{a:[{type:row}]}", ARRAY_OF_JSONOBJECTS,
  *     new CustomComparator(JSONCompareMode.LENIENT, customization));
  * }</pre>
- * 
+ *
  * <p>To verify that the 'id' attribute of every element of array 'a' matches regular expression '\d+'.  This requires a custom comparator to specify regular expression to be used to validate each array element, hence the array of Customization instances:</p>
- * 
+ *
  * <pre>{@code
  * // get length of array we will verify
  * int aLength = ((JSONArray)((JSONObject)JSONParser.parseJSON(ARRAY_OF_JSONOBJECTS)).get("a")).length();
@@ -94,24 +95,24 @@ import org.skyscreamer.jsonassert.comparator.JSONComparator;
  *     new CustomComparator(JSONCompareMode.STRICT_ORDER, new Customization[] { regExArrayValueCustomization });
  * JSONAssert.assertEquals("{a:[{id:X}]}", ARRAY_OF_JSONOBJECTS, regExCustomArrayValueComparator);
  * }</pre>
- * 
+ *
  * <p>To verify that the 'background' attribute of every element of array 'a' alternates between 'white' and 'grey' starting with first element 'background' being 'white':</p>
- * 
+ *
  * <pre>{@code
  * JSONComparator comparator = new DefaultComparator(JSONCompareMode.LENIENT);
  * Customization customization = new Customization("a", new ArrayValueMatcher<Object>(comparator));
  * JSONAssert.assertEquals("{a:[{background:white},{background:grey}]}", ARRAY_OF_JSONOBJECTS,
  *     new CustomComparator(JSONCompareMode.LENIENT, customization));
  * }</pre>
- * 
+ *
  * <p>Assuming JSON to be verified is held in String variable ARRAY_OF_JSONARRAYS and contains:</p>
- * 
+ *
  * <code>{a:[[6,7,8], [9,10,11], [12,13,14], [19,20,21,22]]}</code>
- * 
+ *
  * <p>then:</p>
- * 
+ *
  * <p>To verify that the first three elements of JSON array 'a' are JSON arrays of length 3:</p>
- * 
+ *
  * <pre>{@code
  * JSONComparator comparator = new ArraySizeComparator(JSONCompareMode.STRICT_ORDER);
  * Customization customization = new Customization("a", new ArrayValueMatcher<Object>(comparator, 0, 2));
@@ -119,9 +120,9 @@ import org.skyscreamer.jsonassert.comparator.JSONComparator;
  * }</pre>
  *
  * <p>NOTE: simplified expected JSON strings are not possible in this case as ArraySizeComparator does not support them.</p>
- * 
+ *
  * <p>To verify that the second elements of JSON array 'a' is a JSON array whose first element has the value 9:</p>
- * 
+ *
  * <pre>{@code
  * JSONComparator innerComparator = new DefaultComparator(JSONCompareMode.LENIENT);
  * Customization innerCustomization = new Customization("a[1]", new ArrayValueMatcher<Object>(innerComparator, 0));
@@ -131,9 +132,9 @@ import org.skyscreamer.jsonassert.comparator.JSONComparator;
  * }</pre>
  *
  * <p>To simplify complexity of expected JSON string, the value <code>"{a:[[9]]}"</code> may be replaced by <code>"{a:[9]}"</code> or <code>"{a:9}"</code></p>
- * 
+ *
  * @author Duncan Mackinder
- * 
+ *
  */
 public class ArrayValueMatcher<T> implements LocationAwareValueMatcher<T> {
 	private final JSONComparator comparator;
@@ -144,7 +145,7 @@ public class ArrayValueMatcher<T> implements LocationAwareValueMatcher<T> {
 	 * Create ArrayValueMatcher to match every element in actual array against
 	 * elements taken in sequence from expected array, repeating from start of
 	 * expected array if necessary.
-	 * 
+	 *
 	 * @param comparator
 	 *            comparator to use to compare elements
 	 */
@@ -155,7 +156,7 @@ public class ArrayValueMatcher<T> implements LocationAwareValueMatcher<T> {
 	/**
 	 * Create ArrayValueMatcher to match specified element in actual array
 	 * against first element of expected array.
-	 * 
+	 *
 	 * @param comparator
 	 *            comparator to use to compare elements
 	 * @param index
@@ -169,7 +170,7 @@ public class ArrayValueMatcher<T> implements LocationAwareValueMatcher<T> {
 	 * Create ArrayValueMatcher to match every element in specified range
 	 * (inclusive) from actual array against elements taken in sequence from
 	 * expected array, repeating from start of expected array if necessary.
-	 * 
+	 *
 	 * @param comparator
 	 *            comparator to use to compare elements
 	 * @param from first element in actual array to compared
